Add files via upload
This script predicts the main position of professional league of legends’ players using in-game performance statistics. The models predict with an accuracy of roughly 80%, with AUC scores of above 90%. Different machine learning methodologies are provided together with an aggregation rule for their results. Additionally, I deal with missing observations, provide data visualization and extract insights.