Add files via upload
Thesis for the Master's of Research specialized in Data Science & Statistics.

I evaluate if the modifications implemented in 2017 and 2019 to League of Legends forced the pro players of the strongest region (Korea) to copy the play style of those of a weaker region (Europe). In turn, I argue that this would have improved the competitive balance between the two regions, which could have increased the broadcasting consumption of the European fanbase. The proposed mechanism is novel in both eSports and traditional sports. The evidence supports it for 2019, displaying that Korean players copied Europeans in four of the five different positions of the game. For 2017, only Korean jungle and adc players copied Europeans. These results could have strong implications in the eSports industry, since it would point in the direction of a potential disruptive manipulation of the results of the World Cup, as well as towards possible asymmetric knowledge of eSports owners relevant for betting. Regarding the methodology, I propose a novel framework that evaluates if the main characteristics of two groups converge between periods, and if the characteristics of one group move towards imitating the other. The method is based on a hierarchy of metrics involving shifts in
the characteristics of each group, that allows to distinguish different imitation scenarios. Additionally, it can integrate dimensionality reduction techniques, such as principal components analysis (PCA) to obtain representative group-specific vectors of latent characteristics. Lastly, this is the first example of using PCA to obtain play styles of League of Legends players, which proves to be very insightful.