Add files via upload
Using data that has been gathered in different environments provides new ways of identifying causality, and improved approaches to prediction in targeted scenarios that are heterogeneous to set in which the data is trained. Invariant causal prediction (Peters et al., 2015) identifies causal effects by exploiting the invariance in predictive accuracy of causal models among intervened scenarios. Anchor regression (Rothenhäusler et al., 2018) adds an intermediate minimization of predictive errors whose source is in the heterogeneity of the environment, which improves the predictions of ordinary least squares and the instrumental variables approach for
intermediate ranges of heterogeneity. I develop code in R to implement the latter approach and apply it to a real-world example. A comparison of these results with those of ICP showcases how the latter can dismantle false discoveries of causal effects suggested by the former.